rename base-ws to ws for rails compatibility
