+++ lib/grenache/version.rb
@@ -1,5 +1,5 @@
 module Grenache
   module Websocket
-    VERSION = "0.0.6"
+    VERSION = "0.0.7"
   end
 end
